refactor: extract run_setup() and run_uninstall() into setup.rs
Move setup and uninstall commands from main.rs (2200 -> 2031 lines)
into a dedicated setup.rs module, continuing the line-ratchet effort.
